[refactor] DatabaseCleanUp 및 통합 테스트 support 적용
