+++ src/test/java/com/hhconcert/server/application/UserPointFacadeTest.java
@@ -1,24 +1,19 @@
-package com.hhconcert.server.interfaces.facade;
+package com.hhconcert.server.application;
 
 import com.hhconcert.server.application.dto.PointResult;
 import com.hhconcert.server.application.facade.UserPointFacade;
 import com.hhconcert.server.business.domain.user.entity.User;
+import com.hhconcert.server.config.IntegrationTestSupport;
 import com.hhconcert.server.infrastructure.user.UserJpaRepository;
 import com.hhconcert.server.interfaces.api.point.dto.PointRequest;
 import com.hhconcert.server.interfaces.api.point.dto.UserRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.jdbc.Sql;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@ActiveProfiles("test")
-@SpringBootTest
-@Sql("classpath:test-data.sql")
-class UserPointFacadeTest {
+class UserPointFacadeTest extends IntegrationTestSupport {
 
     @Autowired
     private UserPointFacade userPointFacade;
