+++ src/test/java/com/hhconcert/server/config/DatabaseCleanUp.java
@@ -1,0 +1,41 @@
+package com.hhconcert.server.config;
+
+import jakarta.annotation.PostConstruct;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.PersistenceContext;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Component
+public class DatabaseCleanUp {
+
+    @PersistenceContext
+    private EntityManager entityManager;
+
+    private final List<String> tables = new ArrayList<>();
+
+    @PostConstruct
+    public void afterPropertiesSet() {
+        tables.addAll(
+                entityManager.getMetamodel().getEntities().stream()
+                        .filter(entity -> entity.getJavaType().isAnnotationPresent(Entity.class))
+                        .map(entity -> entity.getName().toLowerCase())
+                        .toList()
+        );
+    }
+
+    @Transactional
+    public void execute() {
+        entityManager.flush();
+        entityManager.createNativeQuery("SET FOREIGN_KEY_CHECKS = 0").executeUpdate();
+        for (String table : tables) {
+            entityManager.createNativeQuery("TRUNCATE TABLE " + table).executeUpdate();
+        }
+        entityManager.createNativeQuery("SET FOREIGN_KEY_CHECKS = 1").executeUpdate();
+    }
+
+}
