+++ src/test/java/com/hhconcert/server/config/IntegrationTestSupport.java
@@ -1,0 +1,21 @@
+package com.hhconcert.server.config;
+
+import com.hhconcert.server.config.DatabaseCleanUp;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+
+@ActiveProfiles("test")
+@SpringBootTest
+public class IntegrationTestSupport {
+
+    @Autowired
+    private DatabaseCleanUp databaseCleanUp;
+
+    @BeforeEach
+    void setUp() {
+        databaseCleanUp.execute();
+    }
+
+}
