+++ src/test/java/com/hhconcert/server/fixture/FacadeTestFixture.java
@@ -1,12 +1,9 @@
-package com.hhconcert.server.interfaces.facade;
+package com.hhconcert.server.fixture;
 
-import com.hhconcert.server.application.dto.ConcertResult;
-import com.hhconcert.server.application.dto.ScheduleResult;
-import com.hhconcert.server.application.dto.SeatResult;
-import com.hhconcert.server.application.facade.ConcertFacade;
 import com.hhconcert.server.business.domain.concert.entity.Concert;
 import com.hhconcert.server.business.domain.payment.entity.Payment;
 import com.hhconcert.server.business.domain.payment.entity.PaymentStatus;
+import com.hhconcert.server.business.domain.queues.entity.Token;
 import com.hhconcert.server.business.domain.reservation.entity.Reservation;
 import com.hhconcert.server.business.domain.reservation.entity.ReservationStatus;
 import com.hhconcert.server.business.domain.schedule.entity.Schedule;
@@ -14,33 +11,19 @@ import com.hhconcert.server.business.domain.seat.entity.Seat;
 import com.hhconcert.server.business.domain.user.entity.User;
 import com.hhconcert.server.infrastructure.concert.ConcertJpaRepository;
 import com.hhconcert.server.infrastructure.payment.PaymentJpaRepository;
+import com.hhconcert.server.infrastructure.queues.TokenJpaRepository;
 import com.hhconcert.server.infrastructure.reservation.ReservationJpaRepository;
 import com.hhconcert.server.infrastructure.schedule.ScheduleJpaRepository;
 import com.hhconcert.server.infrastructure.seat.SeatJpaRepository;
 import com.hhconcert.server.infrastructure.user.UserJpaRepository;
-import com.hhconcert.server.interfaces.api.concert.dto.*;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.jdbc.Sql;
+import org.springframework.stereotype.Component;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.util.List;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.tuple;
-
-@ActiveProfiles("test")
-@SpringBootTest
-@Sql("classpath:test-data.sql")
-class ConcertFacadeTest {
-
-    @Autowired
-    private ConcertFacade concertFacade;
+@Component
+public class FacadeTestFixture {
 
     @Autowired
     private ConcertJpaRepository concertJpaRepository;
@@ -54,15 +37,10 @@ class ConcertFacadeTest {
     private ReservationJpaRepository reservationJpaRepository;
     @Autowired
     private PaymentJpaRepository paymentJpaRepository;
+    @Autowired
+    private TokenJpaRepository tokenJpaRepository;
 
-    LocalDate nowDate;
-    LocalDateTime nowTime;
-
-    @BeforeEach
-    void setUp() {
-        nowDate = LocalDate.now();
-        nowTime = LocalDateTime.now().withNano(0);
-
+    public void concertFixture(LocalDate nowDate, LocalDateTime nowTime) {
         Concert saveConcert1 = concertJpaRepository.save(new Concert("콘서트1", nowDate, nowDate.plusDays(1)));
         Concert saveConcert2 = concertJpaRepository.save(new Concert("콘서트2", nowDate.plusDays(1), nowDate.plusDays(2)));
 
@@ -86,48 +64,47 @@ class ConcertFacadeTest {
         paymentJpaRepository.save(new Payment(user3, saveReserve4, seat3.getPrice(), PaymentStatus.SUCCESS));
     }
 
-    @DisplayName("콘서트 목록을 조회한다.")
-    @Test
-    void getConcerts() {
-        List<ConcertResult> results = concertFacade.getConcerts();
+    public void paymentFixture(LocalDate nowDate, LocalDateTime nowTime) {
+        Concert saveConcert1 = concertJpaRepository.save(new Concert("콘서트1", nowDate, nowDate.plusDays(1)));
+        Concert saveConcert2 = concertJpaRepository.save(new Concert("콘서트2", nowDate.plusDays(1), nowDate.plusDays(2)));
 
-        assertThat(results).extracting("id", "title", "startDate", "endDate")
-                .containsExactly(
-                        tuple(1L, "콘서트1", nowDate, nowDate.plusDays(1)),
-                        tuple(2L, "콘서트2", nowDate.plusDays(1), nowDate.plusDays(2))
-                );
+        Schedule saveSchedule1 = scheduleJpaRepository.save(new Schedule(saveConcert1, nowDate));
+        Schedule saveSchedule2 = scheduleJpaRepository.save(new Schedule(saveConcert1, nowDate.plusDays(1)));
+
+        Seat seat1 = seatJpaRepository.save(new Seat(saveSchedule1, "A1", 75000));
+        Seat seat2 = seatJpaRepository.save(new Seat(saveSchedule1, "B1", 60000));
+        Seat seat3 = seatJpaRepository.save(new Seat(saveSchedule1, "C1", 5000));
+
+        User user1 = userJpaRepository.save(new User("test1234", 80000));
+        User user2 = userJpaRepository.save(new User("qwer1234", 60000));
+        User user3 = userJpaRepository.save(new User("asdf1234", 40000));
+
+        Reservation saveReserve1 = reservationJpaRepository.save(new Reservation(user1, saveConcert1, saveSchedule1, seat1, seat1.getPrice(), ReservationStatus.TEMP, nowTime.plusMinutes(5)));
     }
 
-    @DisplayName("단일 콘서트를 조회한다.")
-    @Test
-    void findConcert() {
-        ConcertResult result = concertFacade.findConcert(new ConcertRequest(1L));
-
-        assertThat(result).extracting("id", "title", "startDate", "endDate")
-                .containsExactly(1L, "콘서트1", nowDate, nowDate.plusDays(1));
+    public void reservationFixture(LocalDate nowDate, LocalDateTime nowTime) {
+        Concert saveConcert1 = concertJpaRepository.save(new Concert("콘서트1", nowDate, nowDate.plusDays(1)));
+        Schedule saveSchedule1 = scheduleJpaRepository.save(new Schedule(saveConcert1, nowDate));
+        Seat seat1 = seatJpaRepository.save(new Seat(saveSchedule1, "A1", 75000));
+        User user1 = userJpaRepository.save(new User("test1234", 80000));
     }
 
-    @DisplayName("특정 콘서트의 예약 날짜를 조회한다.")
-    @Test
-    void getSchedules() {
-        List<ScheduleResult> results = concertFacade.getSchedules(new ConcertRequest(1L));
-
-        assertThat(results).extracting("scheduleId", "date")
-                .containsExactly(
-                        tuple(1L, nowDate),
-                        tuple(2L, nowDate.plusDays(1))
-                );
+    public void userPointFixture() {
+        userJpaRepository.save(new User("test1234", 80000));
+        userJpaRepository.save(new User("asdf1234", 10000));
     }
 
-    @DisplayName("예약 가능 좌석을 조회한다.")
-    @Test
-    void getAvailableSeats() {
-        List<SeatResult> results = concertFacade.getAvailableSeats(new ScheduleRequest(1L));
+    public void queueFixture() {
+        // 사용자 11명 추가
+        for (int i = 0; i < 11; i++) {
+            String userId = "test" + (i + 1);
+            userJpaRepository.save(new User(userId, 10000));
+        }
 
-        assertThat(results).extracting("seatId", "seatNumber" , "price")
-                .containsExactly(
-                        tuple(2L, "B1", 60000)
-                );
+        // 활성화 인원 9명
+        for (int i = 0; i < 9; i++) {
+            String userId = "test" + (i + 1);
+            tokenJpaRepository.save(Token.createForActive(userId));
+        }
     }
-
 }
