--- src/test/resources/test-data.sql
@@ -1,15 +1,0 @@
-drop table if exists concert;
-drop table if exists payment;
-drop table if exists reservation;
-drop table if exists schedule;
-drop table if exists seat;
-drop table if exists token;
-drop table if exists user;
-create table concert (end_date date not null, start_date date not null, concert_id bigint not null auto_increment, created_at datetime(6), modified_at datetime(6), title varchar(255) not null, primary key (concert_id)) engine=InnoDB;
-create table payment (price integer not null, created_at datetime(6), modified_at datetime(6), payment_id bigint not null auto_increment, reserve_id bigint not null, user_seq bigint not null, status enum ('CANCEL','SUCCESS') not null, primary key (payment_id)) engine=InnoDB;
-create table reservation (price integer not null, concert_id bigint not null, created_at datetime(6), expired_at datetime(6), modified_at datetime(6), reserve_id bigint not null auto_increment, schedule_id bigint not null, seat_id bigint not null, user_seq bigint not null, status enum ('CANCEL','COMPLETE','TEMP') not null, primary key (reserve_id)) engine=InnoDB;
-create table schedule (date date, concert_id bigint not null, created_at datetime(6), modified_at datetime(6), schedule_id bigint not null auto_increment, primary key (schedule_id)) engine=InnoDB;
-create table seat (price integer not null, created_at datetime(6), modified_at datetime(6), schedule_id bigint not null, seat_id bigint not null auto_increment, seat_number varchar(255) not null, primary key (seat_id)) engine=InnoDB;
-create table token (active_at datetime(6), created_at datetime(6), expired_at datetime(6), modified_at datetime(6), token_id varchar(255) not null, user_id varchar(255) not null, status enum ('ACTIVE','WAIT') not null, primary key (token_id)) engine=InnoDB;
-create table user (point integer, created_at datetime(6), modified_at datetime(6), user_seq bigint not null auto_increment, user_id varchar(255), primary key (user_seq)) engine=InnoDB;
-alter table payment add constraint UKfg1lnmxqk6w9vpoiivs3urbg1 unique (reserve_id);
